Bumps django-maploom, resolves tag naming conflicts.
The naming conflict was causing Editor to use the Geonode built in version of MapLoom instead of our own. Renaming these tags resolves that conflict and allows us to choose our own MapLoom in the MapStory templates.  Bumped django-maploom because this commit relies on the name changes in that repository as well.